Move maps-api-common to common
