+++ modules/carto/src/api/common.ts
@@ -1,13 +1,4 @@
-import type {SourceOptionalOptions} from './types';
-
-/* eslint-disable camelcase */
-export const SOURCE_DEFAULTS: SourceOptionalOptions = {
-  apiBaseUrl: 'https://gcp-us-east1.api.carto.com',
-  clientId: 'deck-gl-carto',
-  format: 'tilejson',
-  headers: {}
-};
-
+export const DEFAULT_API_BASE_URL = 'https://gcp-us-east1.api.carto.com';
 export const DEFAULT_CLIENT = 'deck-gl-carto';
 export const V3_MINOR_VERSION = '3.2';
 export const MAX_GET_LENGTH = 8192;
